Create input bar component
- Implement each input's features.
- Form submit button enables when all inputs (including custom dropdowns) are filled.
- Initialize input bar states as variables for now.
- Delete unnecessary files.
