+++ src/components/input-bar/index.js
@@ -1,1 +1,2 @@
+import "./InputBar.js";
 import "./DropDown.js";
