--- src/styles/input-bar.css
@@ -1,236 +1,0 @@
-/* .input-bar {
-  width: var(--layout-inner-width);
-  display: flex;
-  justify-content: center;
-  position: absolute;
-  top: -32px;
-  left: 50%;
-  transform: translateX(-50%);
-}
-
-.input-bar > form {
-  height: 76px;
-  padding: 16px 24px;
-  display: flex;
-  background-color: var(--color-white);
-  border: 1px solid var(--color-primary);
-  border-radius: var(--radius-primary);
-  box-shadow: var(--shadow);
-}
-
-.input-bar__item {
-  height: 44px;
-  padding-inline: 24px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  position: relative;
-  border-right: 1px solid var(--color-primary);
-  color: var(--color-primary);
-}
-
-.input-bar__item:first-of-type {
-  padding-left: 0;
-}
-
-.input-bar__item:last-of-type {
-  border-right: none;
-}
-
-.input-bar__item-name {
-  height: 16px;
-  color: var(--color-primary-alt);
-  font-size: var(--fs-bold-regular);
-  font-weight: var(--fw-bold);
-  line-height: var(--lh-bold-regular);
-}
-
-.input-bar__item-input {
-  width: 100%;
-  height: 24px;
-  color: var(--color-primary);
-  font-family: var(--ff-body);
-  font-size: var(--fs-body-regular);
-  font-weight: var(--fw-body);
-  line-height: var(--lh-body-regular);
-  background-color: transparent;
-  border: none;
-}
-
-.input-bar__item-input::placeholder {
-  color: var(--color-primary-alt);
-}
-
-.input-bar__item-input__container {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-} */
-
-/* date input */
-/* #input-bar-date {
-  width: 120px;
-}
-
-.input-bar__item-input#date-input {
-  position: relative;
-}
-
-.input-bar__item-input#date-input::-webkit-inner-spin-button {
-  display: none;
-}
-
-.input-bar__item-input#date-input::-webkit-calendar-picker-indicator {
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  background: transparent;
-  cursor: pointer;
-} */
-
-/* amount input */
-/* #input-bar-amount {
-  width: 175px;
-}
-
-.plus-minus-container {
-  width: 24px;
-  height: 24px;
-}
-
-#is-expense-input,
-#is-expense-input:checked ~ img:nth-of-type(2),
-#is-expense-input:not(:checked) ~ img:nth-of-type(1) {
-  display: none;
-}
-
-#is-expense-input:checked ~ img:nth-of-type(1),
-#is-expense-input:not(:checked) ~ img:nth-of-type(2) {
-  display: block;
-}
-
-#is-expense-input ~ img {
-  cursor: pointer;
-}
-
-.input-bar__item-input#amount-input {
-  width: 78px;
-  text-align: right;
-}
-
-.input-bar__item-input#amount-input::-webkit-inner-spin-button,
-.input-bar__item-input#amount-input::-webkit-outer-spin-button {
-  display: none;
-}
-
-.input-bar__item-input#amount-input::after {
-  content: "원";
-  width: 20px;
-  height: 20px;
-  color: black;
-} */
-
-/* content input */
-/* #input-bar-content {
-  width: 212px;
-} */
-
-/* payment method input */
-/* #input-bar-payment-method {
-  width: 148px;
-} */
-
-/* category input */
-/* #input-bar-category {
-  width: 148px;
-}
-
-.input-bar__item-input__container.dropdown > input {
-  cursor: pointer;
-} */
-
-/* select dropdown */
-/* .input-bar__item .options-container {
-  width: 148px;
-  padding: 0 24px;
-  display: none;
-  position: absolute;
-  top: 65px;
-  left: 0;
-  background-color: var(--color-white);
-  border: 1px solid var(--color-primary);
-  border-radius: var(--radius-primary);
-  box-shadow: var(--shadow);
-  list-style: none;
-}
-
-.input-bar__item .options-container.is-active {
-  display: block;
-}
-
-.input-bar__item .options-container .option {
-  width: 100px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  border-bottom: 1px solid var(--color-primary);
-  cursor: pointer;
-}
-
-.input-bar__item .options-container .option:last-of-type {
-  border-bottom: none;
-}
-
-.input-bar__item .options-container .option span {
-  width: 100%;
-  height: 100%;
-  padding: 16px 0;
-}
-
-.input-bar__item .options-container .option img {
-  filter: var(--filter-secondary-red);
-}
-
-.input-bar__item-input#payment-method-input,
-.input-bar__item-input#category-input {
-  color: var(--color-primary-alt);
-}
-
-#input-bar-payment-method .input-bar__item-input__container,
-#input-bar-category .input-bar__item-input__container {
-  cursor: pointer;
-}
-
-.option-delete-btn {
-  width: 16px;
-  height: 16px;
-  background: transparent;
-  border: none;
-  cursor: pointer;
-}
-
-.option-delete-btn > img {
-  width: 100%;
-  height: 100%;
-} */
-
-/* submit button */
-/* .input-bar > form > button {
-  width: 40px;
-  height: 40px;
-  background-color: var(--color-primary-alt-40);
-  border: none;
-  border-radius: var(--radius-primary);
-  cursor: not-allowed;
-}
-
-.input-bar > form > button:not(button:disabled) {
-  background-color: var(--color-secondary-yellow);
-  cursor: pointer;
-}
-
-.input-bar > form > button > img {
-  width: 24px;
-  height: 24px;
-  filter: var(--filter-white);
-} */
